Third commit

+++ inventryManagementSystem/Models/BillModel.cs
@@ -8,7 +8,7 @@ namespace inventryManagementSystem.Models
 {
     public class BillModel
     {
-        public BillModel(int id, string name, float dose, int days, int amount, float price)
+        public BillModel(int id, string name, float dose, int days, float amount, float price)
         {
             Id = id;
             Name = name;
@@ -22,7 +22,7 @@ namespace inventryManagementSystem.Models
         public string Name { get; set; }
         public float Dose { get; set; }
         public int Days { get; set; }
-        public int Amount { get; set; }
+        public float Amount { get; set; }
         public float Price {  get; set; }
